add net profit logic to invest_in_bond
used f32 instead of u64 for the amount and net_profit in investor
struct. Conversion logic added to the fund function.

when investing in a bond, the investor needs to be added to the
investors vector, or if the investor is already present, the amount
needs to be increased.

Issue: #3